refactor: consolidate pipeline hooks into single usePipelineStatus
